@@ -1,44 +1,46 @@
 {
   "name": "pd-frontend",
-  "version": "0.5.0",
+  "version": "0.6.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "compodoc": "./node_modules/.bin/compodoc -p src/tsconfig.compodoc.json"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.13",
-    "@angular/compiler": "~7.2.15",
-    "@angular/core": "~7.2.13",
-    "@angular/forms": "~7.2.13",
+    "@angular/animations": "^8.0.3",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "~8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "~8.0.3",
+    "@angular/forms": "~8.0.3",
     "@angular/material": "^7.3.7",
     "@angular/material-moment-adapter": "^7.3.7",
-    "@angular/platform-browser": "~7.2.13",
-    "@angular/platform-browser-dynamic": "~7.2.13",
-    "@angular/router": "~7.2.13",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
     "core-js": "^2.5.7",
     "hammerjs": "^2.0.8",
     "ngx-pagination": "^3.2.1",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.2",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "~7.2.13",
+    "@angular-devkit/build-angular": "~0.800.6",
+    "@angular/cli": "^8.0.6",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/language-service": "~8.0.3",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/moment": "^2.13.0",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
+    "compodoc": "0.0.41",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.1",
@@ -49,6 +51,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.1",
     "tslint": "~5.16.0",
-    "typescript": "^3.2.4"
+    "typescript": "3.4.1"
   }
 }
